Add `RuntimeParametersMixin` class
