+++ tests/integration/test_pipe_llms.py
@@ -17,8 +17,9 @@ from typing import Any, Dict, Generator, List
 
 from distilabel.llm.huggingface.transformers import TransformersLLM
 from distilabel.llm.openai import OpenAILLM
+from distilabel.mixins.runtime_parameters import RuntimeParameter
 from distilabel.pipeline.local import Pipeline
-from distilabel.steps.base import RuntimeParameter, Step, StepInput
+from distilabel.steps.base import Step, StepInput
 from distilabel.steps.generators.huggingface import LoadHubDataset
 from distilabel.steps.task.text_generation import TextGeneration
